fix: edit callback session
